requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.2.10
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.1.2
django-allauth==0.45.0
django-appconf==1.0.5
django-cors-headers==3.7.0
django-filter==21.1
django-imagekit==4.1.0
django-js-asset==1.2.2
django-model-utils==4.2.0
django-mptt==0.12.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-ranged-response==0.2.0
django-redis==5.0.0
django-simple-captcha==0.5.14
django-simpleui==2022.1
django-taggit==1.5.0
djangorestframework==3.12.3
djangorestframework-simplejwt==4.7.0
fake-useragent==0.1.11
frozenlist==1.3.0
idna==3.3
jsonfield==3.1.0
lxml==4.7.1
Markdown==3.3.4
multidict==6.0.2
mysqlclient==2.1.0
oauthlib==3.1.1
packaging==21.3
pilkit==2.0
Pillow==8.1.0
pycparser==2.21
PyJWT==2.1.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
redis==4.1.2
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.8
wrapt==1.13.3
yarl==1.7.2
